Add Android port (playable to title) and 64-bit fixes
Vendor SDL2 in android/ and add a Gradle/NDK build that reuses the
SDL2-based PC port. The game now boots to the title screen on arm64
(Redmi 10) with working audio; touch input and packaging come next
(plan phases 4-5).

- android/: SDL2 2.32.x sources + Gradle project + SDLActivity
- pc/CMakeLists.txt: ANDROID branches (shared libmain, GLES loader)
- GLES: shaders #version 300 es; RGBA8 fallback for compressed tex
- Android data paths via SDL_AndroidGetExternalStoragePath() in
  pc_settings/pc_disc/pc_gx_tev; save/card paths under files/
- audio: uintptr_t for OSMesg payloads (system.c, game64, Famicom)
- 64-bit pointer casts across JSystem/emu64/jsyswrapper;
  operator new(size_t); fix TBox copy constructor

+++ src/static/JSystem/JKernel/JKRArchivePub.cpp
@@ -82,7 +82,7 @@ JKRArchive* JKRArchive::mount(const char* path, EMountMode mode, JKRHeap* heap, 
 }
 
 JKRArchive* JKRArchive::mount(void* p1, JKRHeap* heap, EMountDirection mountDirection) {
-    JKRArchive* archive = check_mount_already((s32)p1, heap);
+    JKRArchive* archive = check_mount_already((s32)(intptr_t)p1, heap);
     if (archive != nullptr) {
         return archive;
     }
